feat: enhance CourseSelector with organized course display and access indicators
- Move enhanced course display from SelectionPage to CourseSelector (Course Materials page)
- Remove 'Select a Course' text for cleaner UX
- Implement course-specific access level indicators:
  * Instructor Access (orange) for course admins
  * Student Access (blue) for regular users
- Remove redundant 'Public' badges from course cards
- Add comprehensive documentation in docs/COURSE_SELECTION_ENHANCEMENT.md

Components updated:
- CourseSelector.tsx: Enhanced with organized sections and proper sorting
- CourseCard.tsx: Added access level badges, removed public badge
- CourseSection.tsx: Simplified interface, removed unused props
- SelectionPage.tsx: Removed course display (moved to CourseSelector)

Technical:
- Course-specific permission logic (not global user role)
- Clean, modular component architecture
- Consistent with new sidebar UX/UI design
